kupyna: re-organize modules
