+++ kupyna/src/consts.rs
