+++ kupyna/src/short_compress.rs
@@ -1,10 +1,7 @@
-use crate::tables::{MDS_MATRIX, SBOXES};
+use crate::utils::{add_constant_plus, add_constant_xor, apply_s_box, mix_columns, xor_bytes};
 
 pub(crate) const COLS: usize = 8;
 const ROUNDS: u64 = 10;
-
-const BITS_IN_BYTE: u8 = 8;
-const REDUCTION_POLYNOMIAL: u16 = 0x011d;
 
 type Matrix = [[u8; 8]; 8];
 
@@ -27,11 +24,11 @@ pub(crate) fn compress(prev_vector: &mut [u64; COLS], message_block: &[u8; 64]) 
     }
 }
 
-pub fn t_plus_l(block: [u8; 64]) -> [u8; 64] {
+fn t_plus_l(block: [u8; 64]) -> [u8; 64] {
     let mut state = block_to_matrix(block);
     for nu in 0..ROUNDS {
         state = add_constant_plus(state, nu as usize);
-        state = s_box_layer(state);
+        state = apply_s_box(state);
         state = rotate_rows(state);
         state = mix_columns(state);
     }
@@ -64,37 +61,6 @@ fn matrix_to_block(matrix: Matrix) -> [u8; 64] {
     block
 }
 
-fn add_constant_xor(mut state: Matrix, round: usize) -> Matrix {
-    let rows = 8;
-    for (j, row) in state.iter_mut().enumerate().take(rows) {
-        let constant = ((j * 0x10) ^ round) as u8;
-        row[0] ^= constant;
-    }
-    state
-}
-
-fn add_constant_plus(mut state: Matrix, round: usize) -> Matrix {
-    let rows = 8;
-    for (j, row) in state.iter_mut().enumerate().take(rows) {
-        let mut row_as_u64 = u64::from_le_bytes(row[0..8].try_into().unwrap());
-        row_as_u64 = row_as_u64.wrapping_add(
-            0x00F0F0F0F0F0F0F3u64 ^ (((((rows - j - 1) * 0x10) ^ round) as u64) << 56),
-        );
-        row[0..8].copy_from_slice(&row_as_u64.to_le_bytes());
-    }
-    state
-}
-
-fn s_box_layer(mut state: Matrix) -> Matrix {
-    let cols = 8;
-    for i in 0..cols {
-        for row in state.iter_mut() {
-            row[i] = SBOXES[i % 4][row[i] as usize];
-        }
-    }
-    state
-}
-
 fn rotate_rows(mut state: Matrix) -> Matrix {
     const ROWS: usize = 8;
     let cols = 8;
@@ -117,58 +83,13 @@ fn rotate_rows(mut state: Matrix) -> Matrix {
     state
 }
 
-fn multiply_gf(mut x: u8, mut y: u8) -> u8 {
-    let mut r = 0u8;
-
-    for _ in 0..BITS_IN_BYTE {
-        if y & 1 == 1 {
-            r ^= x;
-        }
-        let hbit = (x & 0x80) >> 7;
-        x <<= 1;
-        if hbit == 1 {
-            x ^= REDUCTION_POLYNOMIAL as u8;
-        }
-        y >>= 1;
-    }
-
-    r
-}
-
-fn mix_columns(state: Matrix) -> Matrix {
-    const ROWS: usize = 8;
-    const COLS: usize = 8;
-
-    let mut result = [[0u8; COLS]; ROWS];
-
-    for col in 0..ROWS {
-        for row in (0..COLS).rev() {
-            let mut product = 0u8;
-            for b in (0..COLS).rev() {
-                product ^= multiply_gf(state[col][b], MDS_MATRIX[row][b]);
-            }
-            result[col][row] = product;
-        }
-    }
-
-    result
-}
-
-pub fn t_xor_l(block: [u8; 64]) -> [u8; 64] {
+pub(crate) fn t_xor_l(block: [u8; 64]) -> [u8; 64] {
     let mut state = block_to_matrix(block);
     for nu in 0..ROUNDS {
         state = add_constant_xor(state, nu as usize);
-        state = s_box_layer(state);
+        state = apply_s_box(state);
         state = rotate_rows(state);
         state = mix_columns(state);
     }
     matrix_to_block(state)
-}
-
-pub(crate) fn xor_bytes(a: [u8; 64], b: [u8; 64]) -> [u8; 64] {
-    let mut result = [0u8; 64];
-    for i in 0..64 {
-        result[i] = a[i] ^ b[i];
-    }
-    result
 }
